pyproject.toml
[tool.poetry]
name = "extraction"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
rdflib = "^7.4.0"
qdrant-client = "^1.15.1"
litellm = "^1.79.1"
fastmcp = "^2.13.0.2"
mcp = "^1.0.0"
strands-agents-tools = "^0.2.14"
strands-agents = {extras = ["otel", "ollama"], version = "^1.15.0"}
langfuse = "^3.9.1"
python-dotenv = "^1.2.1"
streamlit = "^1.51.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
bootstrap_data = "src.bootstrap_data:main"
mcp_server = "src.mcp_server.main:main"
strands = "src.strands.main:main"
ui = "src.ui.runner:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
